+++ ui/src/app/core/components/restart-homebridge/restart-homebridge.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
@@ -7,13 +7,11 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
   templateUrl: './restart-homebridge.component.html',
   styleUrls: ['./restart-homebridge.component.scss'],
 })
-export class RestartHomebridgeComponent implements OnInit {
+export class RestartHomebridgeComponent {
   constructor(
     public activeModal: NgbActiveModal,
     private $router: Router,
   ) { }
-
-  ngOnInit() {}
 
   public onRestartHomebridgeClick() {
     this.$router.navigate(['/restart']);